backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
SQLAlchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.6.1
pydantic-settings==2.1.0
python-multipart==0.0.7
python-dotenv==1.0.1
boto3==1.34.34
Pillow==10.2.0
email-validator>=2.0.0
passlib[bcrypt]>=1.7.4,<2.0
bcrypt>=4.0.0,<5.0.0
